pyproject.toml
[project]
name = "lms"
version = "0.1.0"
description = ""
authors = [
    {name = "Shadyaobuya",email = "shadyaobuyagard@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "strawberry-graphql[debug-server] (>=0.266.0,<0.267.0)",
    "django-allauth[socialaccount] (>=65.7.0,<66.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
